+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This changelog is intended to keep track of my progress and record why I changed
 ## [Unreleased]
 - create healthcheck endpoint
 - handle non existing routes
+- improve common routing to avoid repeated routes
 ## [0.0.3] - 2023-05-06
 - First routed endpoint using standars
 ## [0.0.2] - 2023-04-25